requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
billiard==4.2.1
branca==0.8.1
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
diff-match-patch==20241021
distro==1.9.0
Django==5.1.4
django-celery-beat==2.7.0
django-cors-headers==4.7.0
django-extensions==3.2.3
django-filter==24.3
django-import-export==4.3.5
django-js-asset==3.0.1
django-mptt==0.16.0
django-sslserver==0.22
django-timezone-field==7.1
djangorestframework==3.15.2
flower==1.2.0
folium==0.19.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
kombu==5.4.2
MarkupSafe==3.0.2
numpy==2.2.1
openai==1.59.6
pillow==11.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
qrcode==8.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tablib==3.8.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
xyzservices==2024.9.0
